Bug 2015410 - Part 2: Fix clang-tidy warnings in js/src/builtin/temporal. r=spidermonkey-reviewers,arai
Also fix clang-tidy warnings in "js/src/builtin/temporal".

Differential Revision: https://phabricator.services.mozilla.com/D282345

+++ js/src/builtin/temporal/Temporal.cpp
@@ -598,7 +598,7 @@ static double FractionToDoubleSlow(const T& numerator, const T& denominator) {
     // Move the ignored bits into the proper significand position and adjust the
     // exponent to reflect the now moved out extra bits.
     significand >>= extraBitsCount;
-    exponent += extraBitsCount;
+    exponent += static_cast<int32_t>(extraBitsCount);
 
     MOZ_ASSERT((significand >> SignificandWidthWithImplicitOne) == 0,
                "no excess bits in the significand");
@@ -647,11 +647,11 @@ static double FractionToDoubleSlow(const T& numerator, const T& denominator) {
   if (significandZeros < SignificandLeadingZeros) {
     uint32_t shift = SignificandLeadingZeros - significandZeros;
     significand >>= shift;
-    exponent += shift;
+    exponent += static_cast<int32_t>(shift);
   } else if (significandZeros > SignificandLeadingZeros) {
     uint32_t shift = significandZeros - SignificandLeadingZeros;
     significand <<= shift;
-    exponent -= shift;
+    exponent -= static_cast<int32_t>(shift);
   }
 
   // Combine the individual bits of the double value and return it.
